requirements.txt
bcrypt
pycryptodome
urllib3
Flask
waitress
